+++ packages/utils/src/types/function.ts
@@ -1,1 +1,5 @@
-export type Nullable<T> = T | null
+export type Arrayable<T> = T | T[]
+
+export type Nullable<T> = T | null | undefined
+
+export type Fn<T = void> = () => T