+++ packages/antd/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @raipiot-infra/antd
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @raipiot-infra/utils@0.0.3
+
 ## 0.0.4
 
 ### Patch Changes