+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @raipiot-infra/utils
 
+## 0.0.3
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @raipiot-infra/enums@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes